Rename remaining glossary files to concept
