+++ packages/ndla-ui/src/Concept/component.concept.scss
@@ -1,6 +1,6 @@
 /**
-** GLOSSARY
-** A custom glossary links, show popup with explanation on click
+** CONCEPT
+** A custom concept links, show popup with explanation on click
 **/
 
 .c-concept__item {
